backend/requirements.txt
django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
django-celery-beat==2.8.0
jsonschema==4.23.0
requests==2.32.3
gunicorn==23.0.0
django-redis==5.4.0
